r9035/sofa : Fix windows compil : pow signature problem
Former-commit-id: dd12d018168774f120835d7aa3ab6d348a96a9f8

+++ applications/plugins/image/BranchingImageConverter.h
@@ -143,7 +143,7 @@ protected:
         }
         else
         {
-            unsigned nb = pow( 2, coarseningLevels.getValue() ); // nb fine voxels in a coarse voxel in each direction x/y/z
+            unsigned nb = pow( (float)2, (int)coarseningLevels.getValue() ); // nb fine voxels in a coarse voxel in each direction x/y/z
 
             TransformType::Coord& scale = transform.getScale();
             scale *= nb;
